background logging with stats
Add -l log file option to record most things going to std::cout in a log file. It also supports periodic (every 60 seconds) system stats to be added to the log for FreeBSD. The used sysctl calls should port to Linux with little changes, but I've not test on Linux.